refactor(esde): move ES-DE import into Directories settings
Fold the ES-DE folder-select, run-import, and reset actions into the
Directories settings pane under an "ES-DE Import" section header, and
drop the standalone top-level ES-DE Import menu entry. Rework the
directory list builder to precompute visual rows so both section
headers insert robustly regardless of ROM-folder count.
